refactor: organizing files better
