+++ src/components/Button/stories/Button.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Button } from "./index";
+import { Button } from "../index";
 
 // More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 export default {
